refactor: reorganize utility imports and enhance file handling methods in picker activities
