+++ filepickerlibrary/src/main/java/com/nareshchocha/filepickerlibrary/utilities/extensions/UiExtension.kt
@@ -6,7 +6,7 @@ import android.content.pm.PackageManager
 import android.os.Build
 import androidx.activity.ComponentActivity
 
-fun Context.getRequestedPermissions(): Array<String>? =
+internal fun Context.getRequestedPermissions(): Array<String>? =
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
         packageManager.getPackageInfo(
             packageName,
@@ -19,7 +19,16 @@ fun Context.getRequestedPermissions(): Array<String>? =
         )
     }.requestedPermissions
 
-fun Context.getActivityOrNull(): ComponentActivity? {
+internal fun List<String>.asString() =
+    this
+        .map {
+            it.split(".").lastOrNull() ?: ""
+        }.toString()
+        .let { listString ->
+            listString.substring(1, listString.length - 1).replace(",", " and ")
+        }
+
+internal fun Context.getActivityOrNull(): ComponentActivity? {
     var context = this
     while (context is ContextWrapper) {
         if (context is ComponentActivity) return context
