--- filepickerlibrary/src/test/java/com/nareshchocha/filepickerlibrary/robolectric.properties
@@ -1,2 +1,0 @@
-sdk=28,29
-resourcesMode=binary
